# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是海洋卫士志愿者管理系统的后端API文档",
        "title": "SeaGuard Admin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/volunteers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有志愿者的列表（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "志愿者管理"
                ],
                "summary": "获取志愿者列表",
                "responses": {
                    "200": {
                        "description": "志愿者列表",
                        "schema": {
                            "$ref": "#/definitions/models.VolunteersResponse"
                        }
                    },
                    "403": {
                        "description": "无权限访问",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一个新的志愿者信息（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "志愿者管理"
                ],
                "summary": "创建新志愿者",
                "parameters": [
                    {
                        "description": "志愿者信息",
                        "name": "volunteer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Volunteer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功的志愿者信息",
                        "schema": {
                            "$ref": "#/definitions/models.Volunteer"
                        }
                    },
                    "400": {
                        "description": "请求参数无效",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "无权限访问",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/volunteers/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新指定ID的志愿者信息（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "志愿者管理"
                ],
                "summary": "更新志愿者信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "志愿者ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "志愿者信息",
                        "name": "volunteer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Volunteer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新后的志愿者信息",
                        "schema": {
                            "$ref": "#/definitions/models.Volunteer"
                        }
                    },
                    "400": {
                        "description": "无效的ID参数或请求数据",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "无权限访问",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除指定ID的志愿者（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "志愿者管理"
                ],
                "summary": "删除志愿者",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "志愿者ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "志愿者删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "无效的ID参数",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "无权限访问",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Volunteer": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "integer"
                },
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hours": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.VolunteersResponse": {
            "type": "object",
            "properties": {
                "volunteers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Volunteer"
                    }
                }
            }
        }
    }
}